2.17.2: capped some of traps to not last forever on later depths and made some scale with escalating depths

+++ core/src/main/java/com/shatteredpixel/shatteredpixeldungeon/levels/traps/RockfallTrap.java
@@ -97,7 +97,7 @@ public class RockfallTrap extends Trap {
 			Char ch = Actor.findChar( cell );
 
 			if (ch != null && ch.isAlive()){
-				int damage = Random.NormalIntRange(5+Dungeon.escalatingDepth(), 10+Dungeon.escalatingDepth()*2);
+				long damage = Random.NormalIntRange(5+Dungeon.escalatingDepth(), 10+Dungeon.escalatingDepth()*2);
 				damage -= ch.drRoll();
 				ch.damage( Math.max(damage, 0) , this);
